Fund provider invitations

+++ src/sources/forus-platform/js/angular-1/directives/FundCardSponsorDirective.js
@@ -1,62 +1,68 @@
+let sprintf = require('sprintf-js').sprintf;
+
 let FundCardDirective = function(
     $scope,
     $state,
     FundService,
     ModalService,
-    ProviderFundService
+    ProviderFundService,
+    PushNotificationsService
 ) {
+    let $dir = $scope.$dir = {};
     let topUpInProgress = false;
 
-    $scope.fundCategories = $scope.fund.product_categories.map((val) => {
-        return val.name;
-    });
+    $dir.fund = $scope.fund;
+    $dir.inviteProviders = $scope.inviteProviders || false;
 
-    $scope.changeState = function(state) {
+    $dir.changeState = function(state) {
         FundService.changeState($scope.fund, state).then((res) => {
             $scope.fund = res.data.data;
         });
     };
 
-    $scope.providerApplyFund = function(fund) {
+    $dir.providerApplyFund = function(fund) {
         ProviderFundService.applyForFund(
             $scope.organization.id,
             $scope.fund.id
-        ).then(function(res) {
-            $state.reload();
-        });
+        ).then(() => $state.reload());
     };
 
-    $scope.topUpModal = () => {
-        if (topUpInProgress) return;
+    $dir.topUpModal = () => {
+        if (!topUpInProgress) {
+            topUpInProgress = true;
 
-        topUpInProgress = true;
-
-        ModalService.open('fundTopUp', {
-            fund: $scope.fund
-        }, {
-            onClose: () => topUpInProgress = false
-        });
+            ModalService.open('fundTopUp', {
+                fund: $scope.fund
+            }, {
+                onClose: () => topUpInProgress = false
+            });
+        }
     };
 
-    $scope.deleteFund = function(fund) {
-
+    $dir.deleteFund = function(fund) {
         ModalService.open('modalNotification', {
             type: 'confirm',
             title: 'fund_card_sponsor.confirm_delete.title',
             icon: 'product-error',
             description: 'fund_card_sponsor.confirm_delete.description',
-            confirm: () => {
-                FundService.destroy(
-                    fund.organization_id,
-                    fund.id
-                ).then(function(res) {
-                    $state.reload();
-                });
-            }
+            confirm: () => FundService.destroy(
+                fund.organization_id,
+                fund.id
+            ).then(() => $state.reload())
         });
     };
 
-    $scope.fundOrg = $scope.fund.organization;
+    $dir.inviteProvider = () => {
+        ModalService.open('fundInviteProviders', {
+            fund: $scope.fund,
+            confirm: (res) => {
+                PushNotificationsService.success(
+                    "Providers invited!",
+                    sprintf("%s invitation(s) were sent to providers!", res.length),
+                ) & $state.reload();
+            }
+        });
+    };
 };
 
 module.exports = () => {
@@ -64,7 +70,8 @@ module.exports = () => {
         scope: {
             organization: '=',
             fund: '=',
-            level: '='
+            level: '=',
+            inviteProviders: '<'
         },
         restrict: "EA",
         replace: true,
@@ -74,6 +81,7 @@ module.exports = () => {
             'FundService',
             'ModalService',
             'ProviderFundService',
+            'PushNotificationsService',
             FundCardDirective
         ],
         templateUrl: 'assets/tpl/directives/fund-card-sponsor.html'
